refactor(chat): move ChatPage session identity into a controller hook
The session-identity code in ChatPage.tsx -- the lazy history seed, the
per-mode localStorage slot memory, session tabs, the ?sid deep-link and
URL-sync effects, the mobile drawer's pop bookkeeping, the new-slot
mutation and its retry, slot auto-create, and handleResumeSession --
moves into pages/chat/useChatPageSessionController.ts. The page calls the
hook once where that code used to sit and destructures what it still
reads, so effect registration order is unchanged.

Three edits to the moved text, none of them behaviour: location.key /
location.pathname / location.hash become hook arguments; the dependency
arrays name the refs the hook now receives as arguments (they are
referentially stable, so nothing re-fires); and three comments drop
issue-number references per AGENTS.md.

ChatPage.sessionChipOffline.test.tsx's source contract now reads the
callee declaration from the controller and the chip wiring from the
page.

Second slice of the ChatPage split proposed in #7255.
